requirements.txt
# nothing so far, might add things in the future
# install with `pip install -t lib -r requirements.txt`
google-auth
google-api-python-client
oauth2client
